repositoryのインターフェース
